feat: add side nav layouts for admin hub and supplier portal
- Create src/app/admin/layout.tsx — dark navy sidebar with 11 sections
  (Revenue, Users, Suppliers, Performance, Content, Tickets, Agreements,
  Packages, Audit Logs, Scraper, Review Queue) + user info + sign out
- Create src/app/supplier/layout.tsx — white sidebar with 7 sections
  (Dashboard, Leads, Analytics, Add Product, Billing, Upgrade, Profile)
  + back-to-site link + user info + sign out
- Move scraper from /scraper to /admin/scraper (admin-only); old route
  redirects to new location
- Remove Scraper link from main TopNavbar (now only in admin side nav)
- Strip MainLayout wrapper from supplier/analytics, billing, leads,
  upgrade and admin/scraper-review pages (layout.tsx handles the shell)
